MYR-539 Apply the review's three suggestions
Split internal/arrival/detector.go, which crossed CLAUDE.md's 300-line cap when
this branch gave a ride more than one waypoint to reach. The seam was already
there: detector.go is the subscription, the frame path and the dwell
bookkeeping; detector_advance.go is the single decision they exist to reach.

The completion latch RETURNS a constant instead of `position`. The caller reads
whether a ROW came back, never a value, and returning a column invited a reader
to think some part of the row fed the decision.

A stops edit that arrives with no LegTarget now logs an ERROR instead of falling
out of the switch. The publisher computes it for every stops edit, so reaching
that arm means the invariant broke — and the consequence, a car still driving to
a stop the trip no longer has, is exactly the kind of silence worth refusing.
